Created des-utils.py file and moved all methods that refrence the DES db directly into that file
