+++ easyaccess/eautils/__init__.py
@@ -1,2 +1,2 @@
 __all__ = ["db_api", "des_logo", "dircache", "dtypes",
-           "fileio", "fun_utils", "import_utils", "ea_utils","cli_utils", "db_utils","python_api"]
+           "fileio", "fun_utils", "import_utils", "ea_utils","cli_utils", "db_utils","python_api", "des_utils"]
